Add id to paragraph blocks
The decision to link to the wrapper div instead of the paragraph itself
was made because currently when a link to a paragraph which has
children is taken on Notion, Notion highlights both the paragraph and
its children.
I went back and forth on this decision however so if linking to the
paragraph directly proves more useful we should go for it. Linking to
paragraphs directly feels more intuitive to me anyway