Add the block id as an id attribute to all rendered blocks (#8)
* Add id to headings blocks

* Add id to divider blocks

* Add id to paragraph blocks

The decision to link to the wrapper div instead of the paragraph itself
was made because currently when a link to a paragraph which has
children is taken on Notion, Notion highlights both the paragraph and
its children.
I went back and forth on this decision however so if linking to the
paragraph directly proves more useful we should go for it. Linking to
paragraphs directly feels more intuitive to me anyway

* Add id to quote blocks

* Add id to code blocks

* Add id to lists blocks

The decision to put the id on the list items instead of the lists
themselves is a fairly obvious one to me. No idea how would one put
them on the lists anyway

* Add id to image blocks

I put the id on the figure instead of the img tag because I think they
act as one entity and linking to the whole thing makes most sense
This is as opposed to lists where each item acts independently

* Add id to callout blocks

Same as images here, the whole thing is one unified thing

* Add id to unsupported blocks