Add support for TOML by using out-of-band info
TOML is fairly essential for describing Rust related configuration and
when a website doesn't have highlighting it feels like not enough
effort was put into it. We can't let ourselves be that!
And I am not going to sit tight until Notion adds support for TOML so
I am taking matters into my own hand by adding support for TOML using a
special marker that is unlikely to appear in that specific format
anywhere else